Github PR comment fixes, round 2
