+++ brainscore/metrics/ceiling.py
@@ -162,42 +162,3 @@ ceilings = {
     'cons': InternalConsistency,
     None: NoCeiling,
 }
-
-
-class SplitHalvesConsistencyBaker:
-    # following
-    # https://github.com/brain-score/brain-score/blob/c51b8aa2c94212a9ac56c06c556afad0bb0a3521/brainscore/metrics/ceiling.py#L25-L96
-
-    def __init__(self, num_splits: int, split_coordinate: str, consistency_metric, image_types):
-        """
-        :param num_splits: how many times to create two halves
-        :param split_coordinate: over which coordinate to split the assembly into halves
-        :param consistency_metric: which metric to use to compute the consistency of two halves
-        """
-        self.num_splits = num_splits
-        self.split_coordinate = split_coordinate
-        self.consistency_metric = consistency_metric
-        self.image_types = image_types
-
-    def __call__(self, assembly) -> Score:
-        consistencies, uncorrected_consistencies = [], []
-        splits = range(self.num_splits)
-        for _ in splits:
-            num_subjects = len(set(assembly["subject"].values))
-            half1_subjects = random.sample(range(1, num_subjects), (num_subjects // 2))
-            half1 = assembly[
-                {'presentation': [subject in half1_subjects for subject in assembly['subject'].values]}]
-            half2 = assembly[
-                {'presentation': [subject not in half1_subjects for subject in assembly['subject'].values]}]
-            consistency = self.consistency_metric(half1, half2, image_types=self.image_types, isCeiling=True)
-            uncorrected_consistencies.append(consistency)
-            # Spearman-Brown correction for sub-sampling
-            corrected_consistency = 2 * consistency / (1 + (2 - 1) * consistency)
-            consistencies.append(corrected_consistency)
-        consistencies = Score(consistencies, coords={'split': splits}, dims=['split'])
-        uncorrected_consistencies = Score(uncorrected_consistencies, coords={'split': splits}, dims=['split'])
-        average_consistency = consistencies.median('split')
-        average_consistency.attrs['raw'] = consistencies
-        average_consistency.attrs['uncorrected_consistencies'] = uncorrected_consistencies
-        ceiling_error = np.std(consistencies)
-        return average_consistency, ceiling_error
